feat: Restart Vite on all frontmatter changes
