+++ packages/zudoku/src/vite/plugin.ts
@@ -10,8 +10,8 @@ import { createConfigReloadPlugin } from "./plugin-config-reload.js";
 import viteConfigPlugin from "./plugin-config.js";
 import viteCustomCss from "./plugin-custom-css.js";
 import viteDocsPlugin from "./plugin-docs.js";
+import { viteFrontmatterPlugin } from "./plugin-frontmatter.js";
 import { viteHtmlTransform } from "./plugin-html-transform.js";
-import { viteIconsPlugin } from "./plugin-icons.js";
 import viteMdxPlugin from "./plugin-mdx.js";
 import viteRedirectPlugin from "./plugin-redirect.js";
 import { viteSidebarPlugin } from "./plugin-sidebar.js";
@@ -32,7 +32,7 @@ export default function vitePlugin(
     viteApiKeysPlugin(getCurrentConfig),
     viteAuthPlugin(getCurrentConfig),
     viteDocsPlugin(getCurrentConfig),
-    viteIconsPlugin(),
+    viteFrontmatterPlugin(),
     viteSidebarPlugin(getCurrentConfig),
     viteApiPlugin(getCurrentConfig),
     viteAliasPlugin(getCurrentConfig),
